backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-adk>=1.0.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
python-dotenv>=1.0.0
websockets>=14.0
